requirements.txt
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
arxiv==1.4.8
asgiref==3.8.1
binaryornot==0.4.4
cffi==1.16.0
charset-normalizer==3.3.2
cookiecutter==2.6.0
Django==5.0.4
djangorestframework==3.15.1
exceptiongroup==1.2.0
fastapi==0.110.1
feedparser==6.0.10
flet==0.21.2
flet-core==0.21.2
flet-runtime==0.21.2
fsspec==2024.3.1
httpcore==1.0.5
httptools==0.2.0
httpx==0.27.0
Jinja2==3.1.3
llvmlite==0.42.0
markdown-it-py==3.0.0
mdurl==0.1.2
mistletoe==1.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
odp2md==0.3.3
openai-whisper @ git+https://github.com/openai/whisper@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
qrcode==7.4.2
regex==2023.12.25
repath==0.9.0
requests==2.31.0
rich==13.7.1
scipy==1.13.0
sgmllib3k==1.0.0
sniffio==1.3.1
sounddevice==0.4.6
sqlparse==0.4.4
starlette==0.37.2
sympy==1.12
text-unidecode==1.3
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
triton==2.2.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
uvloop==0.19.0
watchdog==4.0.0
watchgod==0.8.2
websockets==12.0
whisper-app @ git+https://github.com/racoci/whisper-app.git@e912fdb88f4c4b540efaf53ac3286448b1834f9d
